Move notifications into separate file
