+++ Source/Notifications/Notification.Name+ManagedObjectObservation.swift
@@ -1,0 +1,41 @@
+//
+// Wire
+// Copyright (C) 2020 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+
+extension Notification.Name {
+
+    static let ConversationChange = Notification.Name("ZMConversationChangedNotification")
+    static let MessageChange = Notification.Name("ZMMessageChangedNotification")
+    static let UserChange = Notification.Name("ZMUserChangedNotification")
+    static let SearchUserChange = Notification.Name("ZMSearchUserChangedNotification")
+    static let ConnectionChange = Notification.Name("ZMConnectionChangeNotification")
+    static let UserClientChange = Notification.Name("ZMUserClientChangeNotification")
+    static let NewUnreadMessage = Notification.Name("ZMNewUnreadMessageNotification")
+    static let NewUnreadKnock = Notification.Name("ZMNewUnreadKnockNotification")
+    static let NewUnreadUnsentMessage = Notification.Name("ZMNewUnreadUnsentMessageNotification")
+    static let VoiceChannelStateChange = Notification.Name("ZMVoiceChannelStateChangeNotification")
+    static let VoiceChannelParticipantStateChange = Notification.Name("ZMVoiceChannelParticipantStateChangeNotification")
+    static let TeamChange = Notification.Name("TeamChangeNotification")
+    static let LabelChange = Notification.Name("LabelChangeNotification")
+    static let ParticipantRoleChange = Notification.Name("ParticipantRoleChange")
+
+    public static let NonCoreDataChangeInManagedObject = Notification.Name("NonCoreDataChangeInManagedObject")
+
+}
